requirements.txt
folium>=0.20,<1.0
matplotlib>=3.10,<4.0
pandas>=3.0,<4.0
Pillow>=12.0,<13.0
PyMuPDF>=1.27,<2.0
pypdf>=6.10,<7.0
pyshp>=3.0,<4.0
requests>=2.33,<3.0
shapely>=2.1,<3.0
